requirements.txt
streamlit==1.39.0 

ultralytics==8.3.15 

pillow==10.4.0

 numpy==1.26.4 

torch==2.4.1

 torchvision==0.19.1